feat(other): :bricks: change demo from cra to next
