+++ .prettierignore
@@ -1,3 +1,6 @@
+.next
+.vercel
+.tsbuildinfo
 coverage
 dist
 node_modules
