--- demo/config-overrides.js
@@ -1,10 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const ModuleScopePlugin = require('react-dev-utils/ModuleScopePlugin');
-
-module.exports = function override(config) {
-  config.resolve.plugins = config.resolve.plugins.filter(
-    (plugin) => !(plugin instanceof ModuleScopePlugin)
-  );
-
-  return config;
-};
