+++ demo/package.json
@@ -2,30 +2,22 @@
   "name": "demo",
   "private": true,
   "scripts": {
-    "start": "react-app-rewired start"
+    "dev": "next dev",
+    "lint": "next lint"
+  },
+  "dependencies": {
+    "next": "12.1.6",
+    "random-number": "0.0.9",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "recharts": "2.1.10"
   },
   "devDependencies": {
     "@types/random-number": "*",
     "@types/react": "18.0.9",
     "@types/react-dom": "18.0.5",
-    "random-number": "0.0.9",
-    "react": "18.1.0",
-    "react-app-rewired": "2.2.1",
-    "react-dom": "18.1.0",
-    "react-scripts": "5.0.1",
-    "recharts": "2.1.10",
+    "eslint": "8.16.0",
+    "eslint-config-next": "12.1.6",
     "typescript": "4.6.4"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
   }
 }
