--- demo/public/manifest.json
@@ -1,8 +1,0 @@
-{
-  "short_name": "recharts-to-png demo",
-  "name": "recharts-to-png demo",
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
