--- demo/src/index.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-import { createRoot } from 'react-dom/client';
-import App from './App';
-
-const container = document.getElementById('root');
-
-if (container) createRoot(container).render(<App />);
