+++ lib/rollup.config.js
@@ -8,7 +8,7 @@ const external = [
   /@babel\/runtime/,
   'react/jsx-runtime',
   'html2canvas',
-  'react'
+  'react',
 ];
 
 const input = './src/index.ts';
